skills/memos/pyproject.toml
[project]
name = "skill-memos"
version = "0.1.0"
description = "Standalone Memos API client skill script"
readme = "SKILL.md"
requires-python = ">=3.10"
dependencies = [
	"python-dotenv>=1.0.0",
]

[tool.uv]
package = false
